3.1-drf-intro/smart_home/requirements.txt
django
djangorestframework
psycopg2-binary
pillow
